Move LogConfigWatcher to a separate module.
